+++ src/middleware.rs
@@ -1,29 +1,34 @@
 use crate::channel::Sender;
+use crate::helper::load_body;
 use crate::requests::{Request, ShareRequest};
-use actix_http::{error::PayloadError, httpmessage::HttpMessage, Payload};
+use actix_http::{httpmessage::HttpMessage, Payload};
 use actix_service::{Service, Transform};
 use actix_web::{
     dev::{ServiceRequest, ServiceResponse},
     Error,
 };
 use futures::{
-    future::{ok, FutureResult},
-    stream, Future, Poll, Stream,
+    future::{ok, FutureExt, LocalBoxFuture, Ready},
+    stream,
 };
-use std::{cell::RefCell, collections::HashMap, rc::Rc};
+use std::{
+    cell::RefCell,
+    collections::HashMap,
+    rc::Rc,
+    task::{Context, Poll},
+};
 
 impl<S: 'static, B> Transform<S> for ShareRequest
 where
     S: Service<Request = ServiceRequest, Response = ServiceResponse<B>, Error = Error>,
     S::Future: 'static,
-    B: 'static,
 {
     type Request = ServiceRequest;
     type Response = ServiceResponse<B>;
     type Error = Error;
     type InitError = ();
     type Transform = ShareRequestMiddleware<S>;
-    type Future = FutureResult<Self::Transform, Self::InitError>;
+    type Future = Ready<Result<Self::Transform, Self::InitError>>;
 
     fn new_transform(&self, service: S) -> Self::Future {
         ok(ShareRequestMiddleware {
@@ -42,15 +47,14 @@ impl<S, B> Service for ShareRequestMiddleware<S>
 where
     S: Service<Request = ServiceRequest, Response = ServiceResponse<B>, Error = Error> + 'static,
     S::Future: 'static,
-    B: 'static,
 {
     type Request = ServiceRequest;
     type Response = ServiceResponse<B>;
     type Error = Error;
-    type Future = Box<dyn Future<Item = Self::Response, Error = Self::Error>>;
+    type Future = LocalBoxFuture<'static, Result<Self::Response, Self::Error>>;
 
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        self.service.poll_ready()
+    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.service.poll_ready(cx)
     }
 
     fn call(&mut self, mut req: ServiceRequest) -> Self::Future {
@@ -62,28 +66,21 @@ where
         let method = req.method().to_string();
         let path = req.path().to_string();
 
-        Box::new(
-            req.take_payload()
-                .fold(bytes::BytesMut::new(), move |mut body, chunk| {
-                    body.extend_from_slice(&chunk);
-                    Ok::<_, PayloadError>(body)
-                })
-                .map_err(|e| e.into())
-                .and_then(move |bytes| {
-                    let body = bytes.freeze();
-                    let _ = tx.send(Request {
-                        body: String::from_utf8_lossy(&body.to_vec()).to_string(),
-                        headers: req.headers().clone(),
-                        method,
-                        path,
-                        query,
-                    });
+        async move {
+            let body = load_body(req.take_payload()).await?.freeze();
+            let _ = tx.send(Request {
+                body: String::from_utf8_lossy(&body.to_vec()).to_string(),
+                headers: req.headers().clone(),
+                method,
+                path,
+                query,
+            });
 
-                    req.set_payload(Payload::Stream(Box::new(stream::once(Ok(body)))));
+            req.set_payload(Payload::Stream(Box::pin(stream::once(ok(body)))));
 
-                    svc.call(req).and_then(Ok)
-                }),
-        )
+            Ok(svc.call(req).await?)
+        }
+            .boxed_local()
     }
 }
 
@@ -100,7 +97,6 @@ fn extract_query(query: &str) -> HashMap<String, String> {
 
 #[cfg(test)]
 mod tests {
-
     use super::*;
     use actix_web::{
         test::{call_service, init_service, read_body, TestRequest},
@@ -108,21 +104,22 @@ mod tests {
         App, HttpResponse,
     };
 
-    #[test]
-    fn test_middleware() -> Result<(), Error> {
+    #[actix_rt::test]
+    async fn test_middleware() -> Result<(), Error> {
         let (tx, rx) = crate::channel::unbounded();
 
         let mut app =
             init_service(App::new().wrap(ShareRequest { tx }).default_service(
                 route().to(|payload: Payload| HttpResponse::Ok().streaming(payload)),
-            ));
+            ))
+            .await;
 
         let payload = "hello world";
 
         let req = TestRequest::default().set_payload(payload).to_request();
-        let res = call_service(&mut app, req);
+        let res = call_service(&mut app, req).await;
 
-        assert_eq!(read_body(res), payload);
+        assert_eq!(read_body(res).await, payload);
 
         assert_eq!(rx.len(), 1);
         assert_eq!(rx.recv().unwrap().body, payload);